requirements.txt
blinker==1.9.0
click>=8.1.3
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
openai==0.28.1
google-generativeai==0.3.2
Pillow==10.0.0
requests==2.31.0
gTTS==2.3.2

